[release] Always use Clear Linux as upstream for now
The files we with upstream information are not being added yet to the
"www" data, so not available to use. And more importantly there is
still not clarity on how to handle certain cases in that scenario,
e.g. what if we want to change the upstream URL for a new release?

So fallback to use upstream as always Clear Linux for now, and we add
the capability of different upstreams on top.

The upstream information will be generated as part of
https://github.com/clearlinux/mixer-tools/pull/31.

This patch was based on patch by Murilo.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>