[release] Save the list of custom packages
During the mixing process, save the list of custom packages (downstream
packages) that were added on the release candidate and make it available
as part of the update content for that given version. This allows us to
check if a new release is necessary due to changes on the custom
content.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>