[watcher] Call the release job if one is needed
This makes the watcher job actually trigger the release job if it thinks
a new release is necessary. We use the return status to decide that and
also to visually inform if a given execution of this job triggered a new
release or not.

When our stream is up-to-date, job exits with 'success'. If a new release
is needed, the job exits as 'unstable' and call the release job. On
errors, it exits on 'failure'.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>