[watcher] Fetch downstream version number
Adds code to parse Downstream version number so its possible to
compare it against Clear Linux version and decide if a new release is
needed (right now it just outputs a message about the decision).

Note it does not check if the downstream content was updated yet. This
will be done on a later patch.

Also, add a check to verify if downstream version number has at least
the minimum size of characters required: 4.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>