[mixer] Remove bogus code
We have a silly conditional to check if the mixer workspace was already
created to avoid to call mixer init again and "manually" set the version
files. But the problem is that the path used to check the version file
was wrong so we were always calling 'mixer init' anyway.

Since that specific call to mixer init only does what we want (update
the version files), lets use it always and have a simpler code.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>