[mixer] Update upstream bundles clean-up
from mixer 4.1.3 onward, mixer does a good job maintained its upstream
bundles folder, so is no longer necessary to clean it.

However, the bundles list file is reused as is if found in the
workspace and would require manually removal of bundles that happened
to be removed from upstream between two mixer runs, otherwise any
'bundle add' issues an error as well 'build'.

Is easier to just generate the correct list for a given version than
trying to re-use the old list.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>