[release] Annotate the version of used tools
On a downstream workflow, we don't own all the required tools used for
the release process in a way we can just tag its repositories. So, to
achieve reproducible builds, we need to annotate all this information in
a file that will be part of given release meta data. This should allow
us to fully replicate the build environment of a given release.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>