[release] Tag Downstream 'bundles repository'
A given downstream 'bundles repository' may be used across some downstream
instances. Due to this fact, it is necessary to prepend the 'namespace'
(fallback to 'downstream name') on the tag name.

Also, keeping the workaround for mixer 4.3.3 that dont deal well with
hidden folder on the bundle-dir until we enforce a given version of
'mixer'.

Plus: Changed stage.sh to just load and use MIX_VERSION instead of
fetching the version again.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>