[images] Build images in parallel
Major change was on how to properly log and report those building. If
any of the images fail to be created, the stage is flagged as failed and
ister output is also printed. Otherwise ister log is omitted and a 'OK!'
is printed for each images created.

'ister.py' itself relies on 'swupd' that does parallel work, 'xz'
compression run in  parallel as well, so the build of a single image
already makes usage of some parallelism. To avoid overloading the
server, the spawning of jobs is kept at 1 job for each 8 cores in the
server.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>