[images] Add support to create multiple images
Image template files are expected to be found in the 'workflow-repo'
(aka config repo) inside the folder 'images'. Image template files
should have a single word as file name plus the '.json' extension.

The word will be appended on the file name of the final image file and
should try to express the purpose/usage of that image.

If either there is no such folder or no '.json' files are found, the
stage is skipped. All '.json' files inside that folder are assumed to be
image template files appropriated to be used with 'ister.py'.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>