[mixer] Fix support to 'silly' format bumps
Downstream format bumps support dates from the time format bumps were
used by mainline as a form of of the modern 'min version'. Being
intended only to save storage space on the infrastructure servers
instead of actual format changes.

Along the course of this project, the code to handle this endup broken
without notice (mainly because it wasn't execercised anymore due to min
version being used on mainline).

This patch, for historical reasons, fix that support in a way it is
usuablle again for those 'silly' format bumps, even knowing that will be
removed on upcoming commits in favor of true format bumps support. But
at least we can have a working version of this code logged at the
history of this project.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>