[content] Handle the creation of the package list
This is to allow the use of multiple content providers. This silly first
version just concatenates all package lists found in a single file that
will be published, but already stablishes a way for content providers to
add this information.

Each content provider should create a file on
${WORK_DIR}/${PKG_LIST_TMP}<provider name> with the plain list of
packages nvr's to be consumed by this process.

In the future, we will need to address duplicates, sorting and version
mismatch to reflect the final state found on the content repo used by
mixer and published with the release.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>