[mixer] Update Format Bump process
This aligns downstream code with the reviewed version that mainline is
using. It successfully tracks mainline format bumps and works just fine.
But there is a chance that on a near future we will have to touch this
code again if mixer releases a new process specific for downstream
environments.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>