Adding 'make check'
We will use 'shellcheck' to help us to assert code quality. I does
throws some false positive due to how we do inclusion (sourcing) and by
treating our "header" scripts as if they were meant to be executed by
itself, but we can handle this cases by either adding an exclude to the
makefile or adding an exclude directive to the file itself.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>