[release:mixer] Update mixer calls to handle format bump
With the release of mixer 5.7.1, a flag was added to allow for
downstreams to execute format bumps without relying on mixer built-in
format-bump code. This is required for the proper handling of bundle
deletion and to follow the BKM for format bumps suggested by mixer team
itself: afb.sh on mixer-tools official repo.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>