[logging] Make run_and_log blocking
When using the run_and_log command as a part of another script,
operations need to finish before moving on.  Making this a blocking
command.

Opting to have run and log execute a complex command verbatim without
resorting to eval or quoting tricks when passing the command as a
parameter.  As a result, need to supply additional log name information
as a variable for the run_and_log command to align with existing usage.

Update documentation to mention operations are also done with stderr and
better explain default behaviors.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>
Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>