[travis] Check 'make release' via Clear Linux container
Add a Dockerfile for a Clear Linux container which the require tools
and an appropriate user to be build by Travis-CI.

This allow us to validate 'make release' as well. CI will run 'make
release' twice to make sure code is working as intended for subsequent
releases since code path is different.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>