[images] Use 'clr-installer' to create images
'ister.py' is being deprecated in favor of 'clr-installer' as image
generation tool.

'clr-installer' offers a helper command that translates old ister
templates in clr-installer templates, so it should be straight forward
for users of Clear Linux Distro Factory to migrate their configurations
when moving to a new release.

Therefore, there is no need for a transition period where both tools are
supported, so 'ister.py' support is completely removed by this patch.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>