Allow upstream bundles to be built
When an upstream builds a mix, they will source all bundle definitions
locally.  Because upstream logic is largely different from downstream
logic when using mixer, introduce a flag that toggles the commands that
get run if one is an upstram vs a downstream.  If this flag is not
defined, the condition evaluates to false and the downstream logic is
maintained by default.

Also introduce a MIXER_OPTS variable so that upstream may be able to
define in its config that it also wants the offline option, but default
to the native option to preserve existing deployment logic.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>