Add user-configurable global options for mixer commands
Whether the user is an upstream or a downstream determines the global
options that need to be applied to mixer.  Make this something that is
configurable via a variable that can be overridden by a definition in a
configuration repository.

While all global options are not used in all mixer commands, they are
global by definition and we can expect mixer to honor the contract these
commands provide to us with consistent behavior.  An upstream will
always want to use the `--offline` and `--native` flags everywhere
whereas a downstream will always want to use the `--native` flag
everywhere.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>