Infer MIXER_OPTS automatically
MIXER_OPTS should not be a user-configurable parameter because the
options passed to mixer are pre-determined if the user is an upstream or
a downstream.  If a user is an upstream, they will need both `--offline`
and `--native` whereas if a user is a downstream, they will need only
`--native`.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>