[images] Make successfully creating an image optional
The image stage itself is already optional, i.e. It only gets executed
if there are image definitions in the configuration. But once there are
images to be created, current behavior is to successfully build all the
images in order to consider that stage successful as well.

This patch changes that behavior so one or another image failing to
build is okay since they are considered complimentary artifacts for a
release.

Signed-off-by: Vinay Potluri <vinay.potluri@intel.com>