[mixer]: Update how variables are used for mixer.sh
Variables should generally live in globals.sh with sane defaults, and be
overridden by a configuration when they need to change.  Make sure these
variables also have documentation with them.

When supplying defaults for mixer, we want to drop usage of the
`--native` flag because this is the default and only mode for mixer 6.0.
A global MIXER_OPTS with a default value of nothing makes sense to
indicate to the user that this may be overriden.  It would make it
easier in the case when mixer perhaps one day supports a verbose option
for debugging.

Any logic influencing values of variables should be placed in the script
that required that logic.  In this case, if the configuration is for an
upstream and the `--offline flag` has not been supplied, then supply
this flag.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>