[release] Provide mixer with a rpm folder
Do not provide mixer with a rpm repo.  Overriding the local repo with
the desired URL produces unintended side effects.  Instead, provide
mixer with a rpm folder and have the rpms added by mixer.  Even when
this happens, mixer does not properly manage the local-yum folder.  rpms
that disappear in future runs of mixer still persist in the local-yum
folder.  This is why we still need to manage this folder and can't let
mixer manage this just yet.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>
Reviewed-by: Murilo Belluzzo <murilo.belluzzo@intel.com>