Use mixer function wrapper when running MCA
The function wrappers for mixer should be put into common.sh so that
they are made available to other scripts.  Need to use a mixer function
wrapper when running MCA, which was forgotten on the previous commits.
There are no more cases after this one where mixer is called without a
function wrapper.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>