Add logic for running mixer as an upstream
Setting up mixer as an upstream is fairly straightforward.  Only need to
define the bundles and the packages, and then a mix may start.

Due to the nature of how `mixer init` works when the `--offline` flag is
passed, which will be the case for upstreams, an if-condition is
required to run `mixer init` only once.

When content is not found as an upstram, this becomes a hard error
because there are no other sources of content.

As an upstream, format bumps are performed on-demand rather than
automatically as they are for downstreams.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>