Add stubs for signing updates and images
The signing solution required for each user of clr-distro-factory will
vary greatly.  A different solution may be needed for signing updates
than from signing images as well.

If a function is defined as a part of the user's configuration
repository, and this function has as a parameter interface the input and
output files as the first and second parameters, respectively, then this
function is used during the build process.  Users may define a
sign_update function as well as a sign_image function.

This abstracts the particular implementation of the signing solution
away from the build process and allows it to be defined entirely as a
configuration.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>