[image] Fix Image file list parsing
'file_list' was being created as an array of a single entry instead of
an entry per image file. Since the only usage was to later undo the
array in order to get an equivalent list for iteration, remove the array
at all and let word splitting to happen for the 'for' statement.

Signed-off-by: Vinay Potluri <vinay.potluri@intel.com>