[mixer] Fix bundle definition publishing
With the introduction of support to custom paths inside the bundles
repository, stagging of bundle definitions and tagging of the bundles
repository for use cases were BUNDLES_REPO_SRC_DIR is not empty were
broken.

This patch modifies the workflow to better handle that case. It will
keep a pristine copy of the bundles repository as fetched and provide a
copy of the bundles files to 'local-bundles'. During stage, it will use
the bundles def patch to provide the bundles def for publishing without
other artifacts from bundles repo (makefiles and so on). It will
properly tag the bundles for registering build env as well.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>