Finish adding upstream support
An upstream will need to perform a format bump on-demand, as opposed to
a downstream who format bumps automatically.  Introduce a global
variable for this and default it to false in globals.sh so that it may
be overridden by a Jenkins job.

Use prologue.sh to generate a mix version and a mix format for
upstreams.  Once this is done, can start mixing right away.  Update
mixer.sh to go straight to mixing for upstreams.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>