[release] Rename it to 'build'
As pre-work required for splitting the building and publishing process
in two independent pipelines/workflows, allowing for an entry point on
which QA can be executed to decide if publishing can happen.

Since this Pipeline would no longer produce a release, but just a build
instead, rename it to reflect that.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>