[build] Update mixer initialization
Previously, it was an aggreement with mixer-tool maintainer back there
that 'mixer init' could be used to sanitize a Mixer workspace folder to
be sure that all mixer related config files where bootstrapped and with
correct default values. If mixer were to find one of its file already in
the folder, it would only touch it if it was looking outdated.

Currently, this is no longer the case. The preferrable method is now to
only call 'mixer init' the first time mixer workspace is created.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>