[build] Invert 'IS_UPSTREAM'
It makes life much easier if invert this flag to signal when a mix is a
downstream than when it is a upstream. The reason is that a downstream
requires more work to be done on preparing a mix than an upstream, so
it makes the code more readable if we add steps based on the flag than
modifing steps to do less (it is not possible to just skip based on the
flag everywhere in the code the way is written at the moment).

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>