[build] Add MIX_FORMAT instead of reusing DS_FORMAT
DS_FORMAT is used to track the current format of the latest **staged**
(implies published) version of an update stream.

It was being reused along the way to track the format of the mixer being
produced because usually the value will be the same for most cases.

However, during the build process the format for the current mix may
increase when format bumps occurs (potentially several times for
downstream mixes). Therefore it is better to track those separately,
so we keep the original value for reports and logging.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>