[common] Add support for 'upstream' versioning scheme
Version number schemes between head and headless update streams are
different in clr-distro-factory.

As it for now, 'clr-distro-factory' was only prepared to deal with
downstream versioning, on which the version number is composed by the
upstream version number followed a 3-digit number. This imposes a
limited (but fair) amount of builds that can produced by a downstream
based on a given version of upstream. Therefore a downstream version can
be decomposed between up and down fields:

VERSION = UP VER + DOWN VER (final 3-digits)

This makes it very straightforward to figure out which upstream version
was used to produce a downstream version and helps with tracking
format-bumps.

Differently, upstream version is just a number. So the code needs to
change to only make usage of UP and DOWN fields when a downstream is
being produced. Those fields will just be either null or unset for
upstream.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>