[mixer] Upstream mix support - mixer init
For the very first mix of a given update stream, 'clr-distro-factory'
will bootstrap the mixer workspace. This step needs to be done a little
different between upstream and downstream update streams.

We want to make sure 'mixer init' won't add default packages and bundles
by passing the 'offline' flag. For using that flag, we need to provide a
"fake" upstream-version, therefore the mix own version was picked to
supply this requirement.