[release] Bugfix: Don't infer the namespace
After splitting the previous 'release' pipeline into 'build' and
'release' (publish) pipelines, there was a leftover in the pipeline file
for 'release' where it tries to infer the namespace it should use (to
collisions in the builder). This should only be done by what is now the
'build' pipeline. The current 'release' pipeline should follow what is
done for the 'watcher' pipeline.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>