[mixer] Inject Swupd_Root.pem when custom signing
When a user is replacing the signing function of mixer with their own
custom implementation, which implies the `--no-signing` option is being
used, the user is also responsible for injecting the Swupd_Root.pem into
the full chroot.  This is required because swupd is hard-coded by
default to look for the update content certificate in
/usr/share/clear/update-ca/Swupd_Root.pem.

Mixer has a not well-documented feature to support adding sideloaded
content that does not come from a package by creating files in the
SERVER_STATE_DIR which the name `<bundle>-extra-files`.  Because the
content certificate is required for swupd, we inject into the
os-core-update bundle.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>